+++ generic/__init__.py
@@ -1,0 +1,1 @@
+from .remote_connection import remote_connection